requirements.txt
Django==4.0
djangorestframework==3.14.0
psycopg2
djangorestframework-simplejwt==5.2.2
python-decouple==3.7